Working Basic login and registration screen.
