+++ nextjs-blog/pages/profile.jsx
@@ -1,0 +1,27 @@
+//for auth
+import React from 'react';
+import { useUser } from '@auth0/nextjs-auth0/client';
+
+export default function Profile() {
+    const { user, error, isLoading } = useUser();
+  
+    if (isLoading) return <div>Loading...</div>;
+    if (error) return <div>{error.message}</div>;
+  
+    if (user) {
+    return (
+      user && (
+        <div>
+          <img src={user.picture} alt={user.name} />
+          <h2>{user.name}</h2>
+          <p>{user.email}</p>
+          <a href="/api/auth/logout">Logout</a>
+        </div>    
+      )
+    );
+    } else {
+        return (
+            <a href="/api/auth/login">Login</a>
+        )
+    }
+  }
